ommlab_service/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
scikit-image
